Update .gitignore, misc.xml and refactor cash_distribution_calculator.py
Added .gitignore and misc.xml files in the .idea directory for project setup. In the cash_distribution_calculator.py file, the input process is simplified for portfolio balance and three-month-rate. Also, variable names in the calculation and printing of cash and risk balances have been improved for better readability and understanding.